% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representations.R
\name{as_edge_list}
\alias{as_edge_list}
\title{Get the edge list representation of a graph as a matrix.}
\usage{
as_edge_list(.g, .use_names = TRUE, .vert_attr_name = NULL, ...)
}
\arguments{
\item{.g}{graph, currently objects of class \code{igraph} or \code{network}.}

\item{.vert_attr_name}{\code{character(1L)}. The name of a vertex attribute in \code{.g}.}
}
\description{
Get the edge list representation of a graph as a matrix.
}
\author{
Brendan Knapp \email{brendan.knapp@nps.edu}
}
