% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corenet.R
\name{calculate_largest_component}
\alias{calculate_largest_component}
\title{Calculate the largest connected component of a network}
\usage{
calculate_largest_component(network_tile)
}
\arguments{
\item{network_tile}{An \code{sf} object representing the network to be analyzed.}
}
\value{
An \code{sfnetwork} object representing the largest connected component of the network.
}
\description{
This function takes a spatial network represented as an \code{sf} object, converts it into a graph format,
identifies all connected components, and extracts the largest one.
}
