% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corenet.R
\name{calculate_paths_from_point_dist}
\alias{calculate_paths_from_point_dist}
\title{Calculate paths from a given point to centroids within a specified distance range}
\usage{
calculate_paths_from_point_dist(
  network,
  point,
  dist = 500,
  centroids,
  shortest = FALSE
)
}
\arguments{
\item{network}{An sfnetwork object representing the network.}

\item{point}{An sf point object from which paths will be calculated.}

\item{dist}{The maximum distance (in meters) to consider for path calculations.}

\item{centroids}{An sf object containing centroids to which paths are calculated.}

\item{shortest}{Logical indicating whether the shortest paths are calculated (TRUE) or weighted paths (FALSE).}
}
\value{
An sf object containing the paths that meet the criteria or NULL if no paths meet the criteria.
}
\description{
This function determines the network paths from a specific point to multiple centroids based on distance thresholds,
optionally returning the shortest paths.
}
