% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{central_leeds_osm}
\alias{central_leeds_osm}
\title{Central Leeds OSM Network}
\format{
An object of class \code{sf} (inherits from \code{data.frame}).
}
\description{
See the \code{data-raw} folder for the code used to generate this data set.
}
\examples{
head(central_leeds_osm)
library(sf) # for plotting
plot(central_leeds_osm$geometry)
}
\keyword{datasets}
