% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corenet.R
\name{cohesive_network_prep}
\alias{cohesive_network_prep}
\title{Prepare a cohesive cycling network using NPT data}
\usage{
cohesive_network_prep(
  base_network,
  influence_network,
  target_zone,
  crs = "EPSG:27700",
  key_attribute = "road_function",
  attribute_values = c("A Road", "B Road", "Minor Road"),
  use_stplanr = FALSE
)
}
\arguments{
\item{base_network}{Base road network object from Open Road, class 'sf'.}

\item{influence_network}{NPT road network object, which contains influence factors
like all_fastest_bicycle_go_dutch, class 'sf'.}

\item{target_zone}{Target zone, a polygon of a study area, such as a 3km radius around
Edinburgh city centre, class 'sf'.}

\item{crs}{Coordinate reference system to use, default is "EPSG:27700".}

\item{key_attribute}{The attribute in the network data to filter by and influence the outcome.}

\item{attribute_values}{Values of the key_attribute to retain in the network.}

\item{use_stplanr}{Logical value indicating whether to use the stplanr package for merging}
}
\value{
A list containing two elements: a cohesive network and zone data, both class 'sf'.
}
\description{
This function prepares the base network for generating a cohesive cycling network
using NPT data. It integrates two different road network data sources within a specified
target zone to produce a cohesive cycling network that considers specific road attributes.
}
\examples{
library(sf)
library(zonebuilder)
library(dplyr)
library(tmaptools)

# Load demo data 
os_edinburgh_demo_3km = sf::st_set_crs(os_edinburgh_demo_3km, 27700)
NPT_demo_3km = sf::st_set_crs(NPT_demo_3km, 27700)
base_network = sf::st_transform(os_edinburgh_demo_3km, crs = 27700)
influence_network = sf::st_transform(NPT_demo_3km, crs = 27700)
target_zone = zonebuilder::zb_zone("Edinburgh", n_circles = 3) |> sf::st_transform(crs = "EPSG:27700")        

# Prepare the cohesive network
OS_NPT_demo = cohesive_network_prep( base_network = base_network, 
                                influence_network = influence_network, 
                                target_zone = target_zone, 
                                key_attribute = "road_function", 
                                crs = "EPSG:27700", 
                                attribute_values = c("A Road", "B Road", "Minor Road"))

}
