% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corenet.R
\name{create_coherent_network_PMtiles}
\alias{create_coherent_network_PMtiles}
\title{Create coherent network PMtiles}
\usage{
create_coherent_network_PMtiles(folder_path, city_filename, cohesive_network)
}
\arguments{
\item{folder_path}{The directory path where the files will be saved.}

\item{city_filename}{The base name for the output files, in this case using city name.}

\item{cohesive_network}{An sf object representing the cohesive network.}
}
\value{
The output from the Tippecanoe command as a character vector.
}
\description{
This function generates PMtiles for coherent network GeoJSON file using the Tippecanoe tool.
}
