% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corenet.R
\name{get_edinburgh_6km}
\alias{get_edinburgh_6km}
\title{Get the Edinburgh road network, within radius of 6 km of the center}
\usage{
get_edinburgh_6km()
}
\description{
Get the Edinburgh road network, within radius of 6 km of the center
}
\examples{
get_edinburgh_6km()
}
