% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertex-attributes.R
\name{get_vert_attr}
\alias{get_vert_attr}
\alias{get_vert_attr.igraph}
\alias{get_vert_attr.network}
\title{Get the values of a vertex attribute.}
\usage{
get_vert_attr(.g, .vert_attr_name, ...)

\method{get_vert_attr}{igraph}(.g, .vert_attr_name, ...)

\method{get_vert_attr}{network}(.g, .vert_attr_name, ...)
}
\arguments{
\item{.g}{graph, currently objects of class \code{igraph} or \code{network}.}

\item{.vert_attr_name}{\code{character(1L)}. The name of a vertex attribute in \code{.g}.}

\item{...}{arguments passed on to lower-level functions. Currently unused.}
}
\description{
Get the values of a vertex attribute.
}
\author{
Brendan Knapp \email{brendan.knapp@nps.edu}
}
