% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{is_dir}
\alias{is_dir}
\alias{is_dir.igraph}
\alias{is_dir.network}
\title{Are `.g`'s edges directed?}
\usage{
is_dir(.g)

\method{is_dir}{igraph}(.g)

\method{is_dir}{network}(.g)
}
\arguments{
\item{.g}{graph, currently objects of class `igraph` or `network`.}
}
\description{
Are `.g`'s edges directed?
}
\author{
Brendan Knapp \email{brendan.knapp@nps.edu}
}
