% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{is_directed}
\alias{is_directed}
\alias{is_directed.igraph}
\alias{is_directed.network}
\title{Are `.g`'s edges directed?}
\usage{
is_directed(.g)

\method{is_directed}{igraph}(.g)

\method{is_directed}{network}(.g)
}
\arguments{
\item{.g}{graph, currently objects of class `igraph` or `network`.}
}
\description{
Are `.g`'s edges directed?
}
\author{
Brendan Knapp \email{brendan.knapp@nps.edu}
}
