% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{n_edges}
\alias{n_edges}
\alias{n_edges.igraph}
\title{How many edges are there?}
\usage{
n_edges(.g)

\method{n_edges}{igraph}(.g)
}
\arguments{
\item{.g}{graph, currently objects of class `igraph` or `network`.}
}
\description{
How many edges are there?
}
\author{
Brendan Knapp \email{brendan.knapp@nps.edu}
}
