% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{n_verts}
\alias{n_verts}
\alias{n_verts.igraph}
\alias{n_verts.network}
\title{How many vertices are there?}
\usage{
n_verts(.g)

\method{n_verts}{igraph}(.g)

\method{n_verts}{network}(.g)
}
\arguments{
\item{.g}{graph, currently objects of class \code{igraph} or \code{network}.}
}
\description{
How many vertices are there?
}
\author{
Brendan Knapp \email{brendan.knapp@nps.edu}
}
