% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NPT_demo_3km}
\alias{NPT_demo_3km}
\title{Edinburgh NPT Network - 3km Demo Data}
\format{
An object of class \code{sf} (inherits from \code{data.frame}).
}
\usage{
NPT_demo_3km
}
\description{
This data set contains network data for Edinburgh within a 3km radius, sourced from NPT.
It highlights specific network characteristics important for transportation planning.
}
\examples{
data(NPT_demo_3km)
head(NPT_demo_3km)
}
\keyword{datasets}
