% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NPT_demo_6km}
\alias{NPT_demo_6km}
\title{Edinburgh NPT Network - 6km Demo Data}
\format{
An object of class \code{sf} (inherits from \code{data.frame}).
}
\usage{
NPT_demo_6km
}
\description{
This data set contains network data for Edinburgh within a 6km radius, sourced from NPT.
Useful for analyses that require detailed network data and planning metrics.
}
\examples{
data(NPT_demo_6km)
head(NPT_demo_6km)
}
\keyword{datasets}
