% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{os_edinburgh_demo}
\alias{os_edinburgh_demo}
\title{Data from edinburgh's network from OpenRoads}
\format{
A data frame
}
\description{
Data from edinburgh's network from OpenRoads
}
\examples{
head(os_edinburgh_demo)
}
\keyword{datasets}
