% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{os_edinburgh_demo_3km}
\alias{os_edinburgh_demo_3km}
\title{Edinburgh OpenRoads Network - 3km Demo Data}
\format{
An object of class \code{sf} (inherits from \code{data.frame}).
}
\usage{
os_edinburgh_demo_3km
}
\description{
This data set contains network data for Edinburgh within a 3km radius, sourced from OpenRoads.
It is formatted as a spatial data frame (sf object), suitable for spatial analysis demonstrations.
}
\examples{
data(os_edinburgh_demo_3km)
head(os_edinburgh_demo_3km)
}
\keyword{datasets}
