% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{os_edinburgh_demo_6km}
\alias{os_edinburgh_demo_6km}
\title{Edinburgh OpenRoads Network - 6km Demo Data}
\format{
An object of class \code{sf} (inherits from \code{data.frame}).
}
\usage{
os_edinburgh_demo_6km
}
\description{
This data set contains network data for Edinburgh within a 6km radius, sourced from OpenRoads.
It is formatted as a spatial data frame (sf object), which is commonly used in spatial data analysis.
}
\examples{
data(os_edinburgh_demo_6km)
head(os_edinburgh_demo_6km)
}
\keyword{datasets}
