% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corenet.R
\name{prepare_network}
\alias{prepare_network}
\title{Prepare a network data structure by transforming, scoring, and weighting based on road types and conditions}
\usage{
prepare_network(
  network,
  A_Road = 1,
  B_Road = 1,
  Minor_Road = 1e+07,
  transform_crs = 27700
)
}
\arguments{
\item{network}{An sf object representing a spatial network.}

\item{transform_crs}{Numeric CRS code for coordinate transformation, default is 27700.}

\item{road_scores}{A named list where names are road types and values are scores associated with each road type.
Example usage
road_scores = list("A Road" = 1, "B Road" = 1, "Minor Road" = 100000)}
}
\value{
An sfnetwork object with additional attributes like 'arterialness' and 'weight' based on road conditions.
}
\description{
This function takes a spatial network object, casts it to LINESTRING, converts it to an sfnetwork,
and computes scores based on road conditions and classifications. The transformation ensures that
the network is ready for further analytical processes.
}
