% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corenet.R
\name{removeDangles}
\alias{removeDangles}
\title{Remove dangling segments from a road network}
\usage{
removeDangles(network, tolerance = 0.001)
}
\arguments{
\item{network}{An \code{sf} object of class LINESTRING representing the road network.}

\item{tolerance}{The distance tolerance for identifying isolated endpoints as dangling.}
}
\value{
An \code{sf} object with dangling line segments removed.
}
\description{
This function identifies and removes dangling line segments (lines with one end not connected to another line)
within a specified tolerance.
}
