% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corenet.R
\docType{data}
\name{std_pkgs}
\alias{std_pkgs}
\title{Generate a core network from road network data}
\format{
An object of class \code{character} of length 13.
}
\usage{
std_pkgs
}
\description{
This function returns OSM keys that are relevant for active travel
}
\keyword{datasets}
