% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertex-attributes.R
\name{vert_attr_names}
\alias{vert_attr_names}
\alias{vert_attr_names.igraph}
\alias{vert_attr_names.network}
\title{Get all vertex attribute names.}
\usage{
vert_attr_names(.g, ...)

\method{vert_attr_names}{igraph}(.g, ...)

\method{vert_attr_names}{network}(.g, ...)
}
\arguments{
\item{.g}{graph, currently objects of class \code{igraph} or \code{network}.}

\item{...}{arguments passed on to lower-level functions. Currently unused.}
}
\description{
Get all vertex attribute names.
}
\author{
Brendan Knapp \email{brendan.knapp@nps.edu}
}
