% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corenet.R
\name{calculate_paths_from_point_dist}
\alias{calculate_paths_from_point_dist}
\title{Calculate paths from a given point to centroids within a specified distance range}
\usage{
calculate_paths_from_point_dist(
  network,
  point,
  minDistPts = 2,
  maxDistPts = 1500,
  centroids,
  path_type = "shortest",
  max_path_weight = 10
)
}
\arguments{
\item{network}{An sfnetwork object representing the network.}

\item{point}{An sf or sfc object containing a single point feature from which paths will be calculated.}

\item{minDistPts}{The minimum distance (in meters) to consider for path calculations.}

\item{maxDistPts}{The maximum distance (in meters) to consider for path calculations.}

\item{centroids}{An sf object containing centroids to which paths are calculated.}

\item{path_type}{A character string indicating the type of path calculation: 'shortest', 'all_shortest', or 'all_simple'.}

\item{max_path_weight}{\itemize{
\item 'shortest': Calculates the shortest path considering weights.
- 'all_shortest': Calculates all paths that tie for the shortest distance, considering weights.
- 'all_simple': Calculates all simple paths, ignoring weights.
}}
}
\value{
An sf object containing the paths that meet the criteria or NULL if no paths meet the criteria.
}
\description{
This function determines the network paths from a specific point to multiple centroids based on distance thresholds.
It can compute the shortest, all shortest, or all simple paths depending on the path type specified.
}
