% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corenet.R
\name{coherent_network_group}
\alias{coherent_network_group}
\title{This function generates a coherent network grouped by edge betweenness.}
\usage{
coherent_network_group(CITY, ZONE)
}
\arguments{
\item{CITY}{The road network for the city, expected to be an sf object.}

\item{ZONE}{Combined grid buffer, presumably an area of interest within the city, also an sf object.}
}
\value{
A grouped sf network with ranked groups based on mean potential.
}
\description{
This function generates a coherent network grouped by edge betweenness.
}
