% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corenet.R
\name{cohesive_network_prep}
\alias{cohesive_network_prep}
\title{This function prepare the base netwrok for generating cohesive cycling network using NPT data.}
\usage{
cohesive_network_prep(combined_network_tile, crs = "EPSG:27700", parameters)
}
\arguments{
\item{combined_network_tile}{NPT road network object, class sf.}

\item{crs}{Coordinate reference system to use, default is "EPSG:27700".}

\item{parameters}{A list containing specific parameters like 'coherent_area'.}
}
\value{
A list of two elements: a cohesive network and zone data, both as sf objects.
}
\description{
This function prepare the base netwrok for generating cohesive cycling network using NPT data.
}
