% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corenet.R
\name{corenet}
\alias{corenet}
\title{Generate a base network for developing a cohesive cycling network}
\usage{
corenet(
  combined_network_tile,
  network_tile,
  combined_grid_buffer,
  crs = "EPSG:27700",
  dist = 10
)
}
\arguments{
\item{combined_network_tile}{The NPT network, class sf.}

\item{network_tile}{Spatial object ob.}

\item{combined_grid_buffer}{Additional spatial data, currently unused in the road_function}

\item{crs}{Coordinate reference system for transformation, default is "EPSG:27700".}

\item{dist}{The distance threshold used in path calculations, default is 10.}
}
\value{
A spatial object representing the largest cohesive component of the network without dangles.
}
\description{
This function processes the provided network data to extract and analyze
the most significant routes based on a percentile threshold, performs spatial operations,
clusters data points, and calculates the largest network component without dangles.
}
