% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corenet.R
\name{corenet}
\alias{corenet}
\title{Generate a core network from road network data}
\usage{
corenet(net, tbc)
}
\arguments{
\item{net}{A road network object, of class sf}

\item{tbc}{An object to be determined}
}
\value{
A core network object, of class sf
}
\description{
Description to go here...
}
\examples{
corenet()
}
