% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corenet.R
\docType{data}
\name{os_edinburgh_demo}
\alias{os_edinburgh_demo}
\title{Data from OS Open Roads, Edinburgh, 3 km radius}
\format{
An sf data frame
}
\description{
Data from OS Open Roads, Edinburgh, 3 km radius
}
\examples{
library(sf)
# TODO: fix this
# data(os_edinburgh_demo)
# names(os_edinburgh_demo)
os_edinburgh = get_edinburgh_6km()
head(os_edinburgh)
}
\keyword{datasets}
